+++ CHANGELOG.md
@@ -18,7 +18,7 @@
     - Add interfaces. (#156)   
 - Upgrade `hashStruct` function to use `ethers`. (#157)
 - Deploy Nominal and Boost modules with Hardhat. (#158)
-- Publish coverage report on CI/CD. Refactor Jenkinsfile. (#160)
+- Publish coverage report on CI/CD. Refactor Jenkinsfile. (#160, #199)
 - Init Boost deal structure. (#161)
 - Upgrade `eth_signTypedData` function to use `ethers`. Remove now useless `eth-sig-util`. (#163)
 - Run automatically before commit a prettier on `.ts` and `.sol` staged files. (#162)
@@ -45,7 +45,9 @@
         - scheduler (#173)
         - enclave (#176)
     - Handle callback. (#183)
-- Refactor tests. (#177, #178, #179, #181, #188)
+- Implement `claimBoost`:
+    - Refund requester. (#198)
+- Refactor tests. (#177, #178, #179, #181, #188, #200)
 - Add mocha to use test explorer. (#180)
 
 ## v5.4.2